Modifying the type of id from Integer to Long.

+++ src/main/java/tran/example/weatherforecast/repositories/RoleRepository.java
@@ -7,5 +7,5 @@ import tran.example.weatherforecast.domain.security.Role;
  * An interface which will allow Spring Data JPA to create CRUD method implementations at runtime
  * for the Role object(s).
  */
-public interface RoleRepository extends CrudRepository<Role, Integer> {
+public interface RoleRepository extends CrudRepository<Role, Long> {
 }
